+++ codegen/src/main/scala/akka/grpc/gen/javadsl/Service.scala
@@ -1,4 +1,4 @@
-package akka.http.grpc.javadsl
+package akka.grpc.gen.javadsl
 
 import scala.collection.immutable
 import scala.collection.JavaConverters._
@@ -8,6 +8,7 @@ import com.google.protobuf.Descriptors.{ FileDescriptor, ServiceDescriptor }
 final case class Service(packageName: String, name: String, grpcName: String, methods: immutable.Seq[Method]) {
   def serializers: Set[Serializer] = (methods.map(_.deserializer) ++ methods.map(_.serializer)).toSet
 }
+
 object Service {
   def apply(fileDesc: FileDescriptor, serviceDescriptor: ServiceDescriptor): Service = {
     Service(
