+++ codegen/src/main/scala/akka/grpc/gen/scaladsl/Method.scala
@@ -1,14 +1,15 @@
-package akka.http.grpc.javadsl
+package akka.grpc.gen.scaladsl
 
 import com.google.protobuf.Descriptors.{Descriptor, MethodDescriptor}
-import akka.http.grpc._
+import akka.grpc.gen._
+import com.trueaccord.scalapb.compiler.DescriptorPimps
 
-final case class Method(name: String,
+case class Method(name: String,
                   grpcName: String,
                   inputType: Descriptor,
                   inputStreaming: Boolean,
                   outputType: Descriptor,
-                  outputStreaming: Boolean) {
+                  outputStreaming: Boolean)(implicit ops: DescriptorPimps) {
   import Method._
 
   def deserializer = Serializer(inputType)
@@ -22,8 +23,16 @@ final case class Method(name: String,
     if (outputStreaming) "GrpcMarshalling.marshalStream"
     else "GrpcMarshalling.marshal"
 
+  def parameterType =
+    if (inputStreaming) s"Source[${messageType(inputType)}, NotUsed]"
+    else messageType(inputType)
+
   def inputTypeUnboxed = messageType(inputType)
   def outputTypeUnboxed = messageType(outputType)
+
+  def returnType =
+    if (outputStreaming) s"Source[${messageType(outputType)}, NotUsed]"
+    else s"Future[${messageType(outputType)}]"
 
   val methodType: MethodType = {
     (inputStreaming, outputStreaming) match {
@@ -33,19 +42,11 @@ final case class Method(name: String,
       case (true, true)   => BidiStreaming
     }
   }
-
-  def getParameterType =
-    if (inputStreaming) s"Source<${getMessageType(inputType)}, NotUsed>"
-    else getMessageType(inputType)
-
-  def getReturnType =
-    if (outputStreaming) s"Source<${getMessageType(outputType)}, NotUsed>"
-    else s"CompletionStage<${getMessageType(outputType)}>"
 }
 
 
 object Method {
-  def apply(descriptor: MethodDescriptor): Method = {
+  def apply(descriptor: MethodDescriptor)(implicit ops: DescriptorPimps): Method = {
     Method(
       name = methodName(descriptor.getName),
       grpcName = descriptor.getName,
@@ -59,13 +60,11 @@ object Method {
   private def methodName(name: String) =
     name.head.toLower +: name.tail
 
-  def messageType(t: Descriptor) =
-    "_root_." + t.getFile.getOptions.getJavaPackage + "." + protoName(t) + "." + t.getName
-
-  /** Java API */
-  def getMessageType(t: Descriptor) = {
-    t.getFile.getOptions.getJavaPackage + "." + outerClass(t) + t.getName
+  def messageType(messageType: Descriptor)(implicit ops: DescriptorPimps) = {
+    import ops._
+    messageType.scalaTypeName
   }
+
 
   private def outerClass(t: Descriptor) = {
     if (t.getFile.toProto.getOptions.getJavaMultipleFiles) ""
