+++ codegen/src/main/scala/akka/grpc/gen/scaladsl/ScalaBothCodeGenerator.scala
@@ -1,4 +1,4 @@
-package akka.http.grpc.scaladsl
+package akka.grpc.gen.scaladsl
 
 object ScalaBothCodeGenerator extends ScalaServerCodeGenerator
   with ScalaClientCodeGenerator
