+++ codegen/src/main/twirl/templates/ScalaCommon/ApiTrait.scala.txt
@@ -1,4 +1,4 @@
-@(service: akka.http.grpc.scaladsl.Service)
+@(service: akka.grpc.gen.scaladsl.Service)
 
 package @service.packageName
 
@@ -17,7 +17,7 @@ object @{service.name} {
   val name = "@service.grpcName"
 
   object Serializers {
-    import akka.http.grpc.scaladsl.ScalapbProtobufSerializer
+    import akka.grpc.scaladsl.ScalapbProtobufSerializer
     @for(serializer <- service.serializers) {
     val @serializer.name = @serializer.init
     }
