+++ codegen/src/main/twirl/templates/ScalaServer/Handler.scala.txt
@@ -1,10 +1,10 @@
-@(service: akka.http.grpc.scaladsl.Service)
+@(service: akka.grpc.gen.scaladsl.Service)
 
 package @service.packageName
 
 import scala.concurrent.{ ExecutionContext, Future }
 
-import akka.http.grpc.scaladsl.{ GrpcExceptionHandler, GrpcMarshalling, ScalapbProtobufSerializer }
+import akka.grpc.scaladsl.{ GrpcExceptionHandler, GrpcMarshalling, ScalapbProtobufSerializer }
 
 import akka.http.scaladsl.model.{ HttpRequest, HttpResponse, StatusCodes }
 import akka.http.scaladsl.model.Uri.Path
