+++ runtime/src/main/scala/akka/grpc/javadsl/GoogleProtobufSerializer.scala
@@ -1,6 +1,6 @@
-package akka.http.grpc.javadsl
+package akka.grpc.javadsl
 
-import akka.http.grpc.ProtobufSerializer
+import akka.grpc.ProtobufSerializer
 
 class GoogleProtobufSerializer[T <: com.google.protobuf.Message](clazz: Class[T]) extends ProtobufSerializer[T] {
   override def serialize(t: T) = akka.util.ByteString(t.toByteArray)
