+++ sbt-plugin/src/main/scala/akka/grpc/sbt/AkkaJavaServerGrpcCodeGenerator.scala
@@ -1,6 +1,6 @@
 package akka.grpc.sbt
 
-import akka.http.grpc.javadsl.JavaServerCodeGenerator
+import akka.grpc.gen.javadsl.JavaServerCodeGenerator
 
 /** An easy to use accessor to be used reflectively in the main sbt project */
 class AkkaJavaServerGrpcCodeGenerator {
